# Use batch APIs to create arrays and hashes