Basically a port of https://github.com/ruby/json/pull/678

Rather than to allocate the container and push elements one by one, we accumulate them on a stack and then use the faster batch APIs to directly create the final container.

The original action stack remains, as we need to keep track of what we're currently parsing and how big it is.
    
But for the recursive cases, we no longer need to create a child stack.