@byroot 2024-11-11
Closing in favor of https://github.com/msgpack/msgpack-ruby/pull/370.

I don't think the batch API are worth it here, but it's worth no longer allocating a child stack.